Query settingComponents from craft config instead of Node instance
settingComponents' origin path is `craft.addition`,
which will be copy to attribute `addition` in Node instance.
However, settingComponents is an array of Vue component option which
contain some hook function, so it cannot correctly be serialized to JSON.
In order to implement export functionality,
we change the path of settingComponents config,
and query settingComponents from craft config instead of Node instance.